# Conflicts:
#	README.md
#	pom.xml
#	threatconnect-app/apps-core/pom.xml
#	threatconnect-app/apps-util/pom.xml
#	threatconnect-app/playbooks-core/pom.xml
#	threatconnect-app/playbooks-test/pom.xml
#	threatconnect-app/pom.xml
#	threatconnect-sdk-addons/app-package-plugin/README.md
#	threatconnect-sdk-addons/app-package-plugin/pom.xml
#	threatconnect-sdk-addons/opendxl/pom.xml
#	threatconnect-sdk-addons/pom.xml
#	threatconnect-sdk-addons/stix-reader/pom.xml
#	threatconnect-sdk-addons/stix-writer/pom.xml
#	threatconnect-sdk-addons/threatconnect-parser/README.md
#	threatconnect-sdk-addons/threatconnect-parser/pom.xml
#	threatconnect-sdk/pom.xml
#	threatconnect-sdk/threatconnect-model/pom.xml
#	threatconnect-sdk/threatconnect-sdk-core/README.md
#	threatconnect-sdk/threatconnect-sdk-core/pom.xml

+++ threatconnect-sdk-addons/threatconnect-parser/src/main/java/com/threatconnect/sdk/parser/service/writer/IndicatorWriter.java
@@ -190,7 +190,7 @@ public abstract class IndicatorWriter<E extends Indicator, T extends com.threatc
 		}
 	}
 	
-	public void associateGroup(final GroupType groupType, final Integer savedID, final String ownerName)
+	public void associateGroup(final GroupType groupType, final Long savedID, final String ownerName)
 		throws AssociateFailedException, IOException
 	{
 		try
