+++ threatconnect-sdk/threatconnect-sdk-core/src/main/java/com/threatconnect/sdk/client/fluent/VictimNetworkAccountBuilder.java
@@ -4,14 +4,14 @@ import com.threatconnect.sdk.server.entity.VictimNetworkAccount;
 
 public class VictimNetworkAccountBuilder
 {
-    private Integer id;
+    private Long id;
     private String name;
     private String type;
     private String webLink;
     private String account;
     private String network;
 
-    public VictimNetworkAccountBuilder withId(Integer id)
+    public VictimNetworkAccountBuilder withId(Long id)
     {
         this.id = id;
         return this;
