+++ threatconnect-sdk/threatconnect-sdk-core/src/main/java/com/threatconnect/sdk/client/writer/associate/AbstractAttributeAssociateWriterAdapter.java
@@ -64,12 +64,12 @@ public abstract class AbstractAttributeAssociateWriterAdapter<T,P> extends Abstr
     }
 
     @Override
-    public WriteListResponse<String> addAttributeSecurityLabels(P uniqueId, Integer attributeId, List<String> securityLabels ) throws IOException {
+    public WriteListResponse<String> addAttributeSecurityLabels(P uniqueId, Long attributeId, List<String> securityLabels ) throws IOException {
         return addAttributeSecurityLabels(uniqueId, attributeId, securityLabels, null);
     }
 
     @Override
-    public WriteListResponse<String> addAttributeSecurityLabels(P uniqueId, Integer attributeId, List<String> securityLabels, String ownerName)
+    public WriteListResponse<String> addAttributeSecurityLabels(P uniqueId, Long attributeId, List<String> securityLabels, String ownerName)
         throws IOException {
 
         Map<String, Object> map = createParamMap("id", uniqueId, "attributeId", attributeId);
@@ -80,12 +80,12 @@ public abstract class AbstractAttributeAssociateWriterAdapter<T,P> extends Abstr
     }
 
     @Override
-    public ApiEntitySingleResponse addAttributeSecurityLabel(P uniqueId, Integer attributeId, String securityLabel) throws IOException, FailedResponseException {
+    public ApiEntitySingleResponse addAttributeSecurityLabel(P uniqueId, Long attributeId, String securityLabel) throws IOException, FailedResponseException {
         return addAttributeSecurityLabel(uniqueId, attributeId, securityLabel, null);
     }
 
     @Override
-    public ApiEntitySingleResponse addAttributeSecurityLabel(P uniqueId, Integer attributeId, String securityLabel, String ownerName)
+    public ApiEntitySingleResponse addAttributeSecurityLabel(P uniqueId, Long attributeId, String securityLabel, String ownerName)
         throws IOException, FailedResponseException {
 
         Map<String, Object> map = createParamMap("id", uniqueId, "attributeId", attributeId, "securityLabel", securityLabel);
@@ -104,7 +104,7 @@ public abstract class AbstractAttributeAssociateWriterAdapter<T,P> extends Abstr
     public WriteListResponse<Attribute> updateAttributes(P uniqueId, List<Attribute> attributes, String ownerName) throws IOException {
 
         Map<String, Object> map = createParamMap("id", uniqueId);
-        List<Integer> idList = new ArrayList<>();
+        List<Long> idList = new ArrayList<>();
         for( Attribute a : attributes )     idList.add( a.getId() );
         WriteListResponse data = updateListWithParam(getUrlBasePrefix() + ".byId.attributes.byId", AttributeListResponse.class, ownerName, map, "attributeId", idList, attributes);
 
@@ -127,28 +127,28 @@ public abstract class AbstractAttributeAssociateWriterAdapter<T,P> extends Abstr
     }
 
     @Override
-    public WriteListResponse<Integer> deleteAttributes(P uniqueId, List<Integer> attributes) throws IOException {
+    public WriteListResponse<Long> deleteAttributes(P uniqueId, List<Long> attributes) throws IOException {
         return deleteAttributes(uniqueId, attributes, null);
     }
 
     @Override
-    public WriteListResponse<Integer> deleteAttributes(P uniqueId, List<Integer> attribute, String ownerName) throws IOException {
+    public WriteListResponse<Long> deleteAttributes(P uniqueId, List<Long> attribute, String ownerName) throws IOException {
         Map<String, Object> map = createParamMap("id", uniqueId);
-        List<Integer> idList = new ArrayList<>();
-        for(Integer it : attribute)    idList.add( it );
-        WriteListResponse<Integer> data = deleteList(getUrlBasePrefix() + ".byId.attributes.byId"
+        List<Long> idList = new ArrayList<>();
+        for(Long it : attribute)    idList.add( it );
+        WriteListResponse<Long> data = deleteList(getUrlBasePrefix() + ".byId.attributes.byId"
                 , AttributeResponse.class, ownerName, map, "attributeId", idList);
 
         return data;
     }
 
     @Override
-    public ApiEntitySingleResponse deleteAttribute(P uniqueId, Integer attribute) throws IOException, FailedResponseException {
+    public ApiEntitySingleResponse deleteAttribute(P uniqueId, Long attribute) throws IOException, FailedResponseException {
         return deleteAttribute(uniqueId, attribute, null);
     }
 
     @Override
-    public ApiEntitySingleResponse deleteAttribute(P uniqueId, Integer attribute, String ownerName) throws IOException, FailedResponseException {
+    public ApiEntitySingleResponse deleteAttribute(P uniqueId, Long attribute, String ownerName) throws IOException, FailedResponseException {
         Map<String, Object> map = createParamMap("id", uniqueId, "attributeId", attribute);
         ApiEntitySingleResponse item = deleteItem(getUrlBasePrefix() + ".byId.attributes.byId", AttributeResponse.class, ownerName, map);
 
@@ -156,12 +156,12 @@ public abstract class AbstractAttributeAssociateWriterAdapter<T,P> extends Abstr
     }
 
     @Override
-    public WriteListResponse<String> deleteAttributeSecurityLabels(P uniqueId, Integer attributeId, List<String> securityLabels) throws IOException {
+    public WriteListResponse<String> deleteAttributeSecurityLabels(P uniqueId, Long attributeId, List<String> securityLabels) throws IOException {
         return deleteAttributeSecurityLabels(uniqueId, attributeId, securityLabels, null);
     }
 
     @Override
-    public WriteListResponse<String> deleteAttributeSecurityLabels(P uniqueId, Integer attributeId, List<String> securityLabels, String ownerName) throws IOException {
+    public WriteListResponse<String> deleteAttributeSecurityLabels(P uniqueId, Long attributeId, List<String> securityLabels, String ownerName) throws IOException {
        Map<String, Object> map = createParamMap("id", uniqueId);
        WriteListResponse<String> data = deleteList(getUrlBasePrefix() + ".byId.attributes.byId.securityLabels.byName"
                 , SecurityLabelResponse.class, ownerName, map, "securityLabel", securityLabels);
@@ -170,12 +170,12 @@ public abstract class AbstractAttributeAssociateWriterAdapter<T,P> extends Abstr
     }
 
     @Override
-    public ApiEntitySingleResponse deleteAttributeSecurityLabel(P uniqueId, Integer attributeId, String securityLabel) throws IOException, FailedResponseException {
+    public ApiEntitySingleResponse deleteAttributeSecurityLabel(P uniqueId, Long attributeId, String securityLabel) throws IOException, FailedResponseException {
         return deleteAttributeSecurityLabel(uniqueId, attributeId, securityLabel, null);
     }
 
     @Override
-    public ApiEntitySingleResponse deleteAttributeSecurityLabel(P uniqueId, Integer attributeId, String securityLabel, String ownerName) throws IOException, FailedResponseException {
+    public ApiEntitySingleResponse deleteAttributeSecurityLabel(P uniqueId, Long attributeId, String securityLabel, String ownerName) throws IOException, FailedResponseException {
         Map<String, Object> map = createParamMap("id", uniqueId, "securityLabel", securityLabel);
         ApiEntitySingleResponse item = deleteItem(getUrlBasePrefix() + ".byId.attributes.byId.securityLabels.byName", SecurityLabelResponse.class, ownerName, map);
