+++ threatconnect-sdk/threatconnect-sdk-core/src/main/java/com/threatconnect/sdk/client/writer/associate/AttributeAssociateWritable.java
@@ -23,14 +23,14 @@ public interface AttributeAssociateWritable<P> {
     public ApiEntitySingleResponse addAttribute(P uniqueId, Attribute attribute, String ownerName)
         throws IOException, FailedResponseException;
 
-    public WriteListResponse<String> addAttributeSecurityLabels(P uniqueId, Integer attributeId, List<String> securityLabels) throws IOException;
+    public WriteListResponse<String> addAttributeSecurityLabels(P uniqueId, Long attributeId, List<String> securityLabels) throws IOException;
 
-    public WriteListResponse<String> addAttributeSecurityLabels(P uniqueId, Integer attributeId, List<String> securityLabels, String ownerName)
+    public WriteListResponse<String> addAttributeSecurityLabels(P uniqueId, Long attributeId, List<String> securityLabels, String ownerName)
         throws IOException;
 
-    public ApiEntitySingleResponse addAttributeSecurityLabel(P uniqueId, Integer attributeId, String securityLabel) throws IOException, FailedResponseException;
+    public ApiEntitySingleResponse addAttributeSecurityLabel(P uniqueId, Long attributeId, String securityLabel) throws IOException, FailedResponseException;
 
-    public ApiEntitySingleResponse addAttributeSecurityLabel(P uniqueId, Integer attributeId, String securityLabel, String ownerName)
+    public ApiEntitySingleResponse addAttributeSecurityLabel(P uniqueId, Long attributeId, String securityLabel, String ownerName)
         throws IOException, FailedResponseException;
 
     public WriteListResponse<Attribute> updateAttributes(P uniqueId, List<Attribute> attributes)
@@ -44,25 +44,25 @@ public interface AttributeAssociateWritable<P> {
     public ApiEntitySingleResponse updateAttribute(P uniqueId, Attribute attribute, String ownerName)
         throws IOException, FailedResponseException;
 
-    public WriteListResponse<Integer> deleteAttributes(P uniqueId, List<Integer> attributes)
+    public WriteListResponse<Long> deleteAttributes(P uniqueId, List<Long> attributes)
         throws IOException;
 
-    public WriteListResponse<Integer> deleteAttributes(P uniqueId, List<Integer> attribute, String ownerName)
+    public WriteListResponse<Long> deleteAttributes(P uniqueId, List<Long> attribute, String ownerName)
         throws IOException;
 
-    public ApiEntitySingleResponse deleteAttribute(P uniqueId, Integer attribute) throws IOException, FailedResponseException;
+    public ApiEntitySingleResponse deleteAttribute(P uniqueId, Long attribute) throws IOException, FailedResponseException;
 
-    public ApiEntitySingleResponse deleteAttribute(P uniqueId, Integer attribute, String ownerName)
+    public ApiEntitySingleResponse deleteAttribute(P uniqueId, Long attribute, String ownerName)
         throws IOException, FailedResponseException;
 
-    public WriteListResponse<String> deleteAttributeSecurityLabels(P uniqueId, Integer attributeId, List<String> securityLabels) throws IOException, FailedResponseException;
+    public WriteListResponse<String> deleteAttributeSecurityLabels(P uniqueId, Long attributeId, List<String> securityLabels) throws IOException, FailedResponseException;
 
-    public WriteListResponse<String> deleteAttributeSecurityLabels(P uniqueId, Integer attributeId, List<String> securityLabels, String ownerName)
+    public WriteListResponse<String> deleteAttributeSecurityLabels(P uniqueId, Long attributeId, List<String> securityLabels, String ownerName)
         throws IOException, FailedResponseException;
 
-    public ApiEntitySingleResponse deleteAttributeSecurityLabel(P uniqueId, Integer attributeId, String securityLabel) throws IOException, FailedResponseException;
+    public ApiEntitySingleResponse deleteAttributeSecurityLabel(P uniqueId, Long attributeId, String securityLabel) throws IOException, FailedResponseException;
 
-    public ApiEntitySingleResponse deleteAttributeSecurityLabel(P uniqueId, Integer attributeId, String securityLabel, String ownerName)
+    public ApiEntitySingleResponse deleteAttributeSecurityLabel(P uniqueId, Long attributeId, String securityLabel, String ownerName)
         throws IOException, FailedResponseException;
 
 }
