+++ threatconnect-sdk/threatconnect-sdk-core/src/main/java/com/threatconnect/sdk/server/entity/Incident.java
@@ -32,7 +32,7 @@ public class Incident extends Group
         super();
     }
 
-    public Incident(Integer id, String name, String type, Owner owner, String ownerName, Date dateAdded, String webLink, Date eventDate)
+    public Incident(Long id, String name, String type, Owner owner, String ownerName, Date dateAdded, String webLink, Date eventDate)
     {
         super(id, name, type, owner, ownerName, dateAdded, webLink);
         this.eventDate = eventDate;
